Deleted unused files&Refactor NavBar&Made changes to More&Add redux toggle for More&Update the relative test
