+++ client/src/layout/NavBar.jsx
@@ -1,110 +1,94 @@
 import React from 'react';
-import { withStyles } from '@material-ui/core/styles';
 import { Link } from 'react-router-dom';
-import { BottomNavigation, BottomNavigationAction } from '@material-ui/core/';
+import { makeStyles } from '@material-ui/core/styles';
+import { AppBar, Toolbar, Button, Typography, Grid } from '@material-ui/core';
+
 import { connect } from 'react-redux';
 import FeaturedPlayListOutlinedIcon from '@material-ui/icons/FeaturedPlayListOutlined';
 import StarBorderIcon from '@material-ui/icons/StarBorder';
 import LocationOnOutlinedIcon from '@material-ui/icons/LocationOnOutlined';
+import MoreHorizIcon from '@material-ui/icons/MoreHoriz';
+import More from 'layout/More';
 import PropTypes from 'prop-types';
-import actions from '../redux/actions/actions';
-import More from '../components/more/More';
+import actions from 'redux/actions/actions';
 
-const styles = {
-  root: {
-    width: '100%',
-    background: `linear-gradient(0deg, #4760ff, #82a4f8)`,
-    position: 'fixed',
-    height: '75px',
+const useStyles = makeStyles(theme => ({
+  // making the style of NavBar
+  appBar: {
+    top: 'auto',
     bottom: 0,
+    height: theme.spacing(8),
+    background: `linear-gradient(0deg, #4760ff, #82a4f8)`,
   },
-  actionItem: {
-    color: 'white',
-    transition: 'all .3s ease-in-out',
-    '&:hover': {
-      textDecoration: 'none',
-      color: 'white',
-      transform: 'scale(1.1)',
-    },
-    '&$selected': {
-      color: 'white',
-      backgroundColor: 'red',
-    },
+  toolBar: {
+    justifyContent: 'center',
   },
-  selected: {},
   buttons: {
-    fontSize: '40px',
-
     transition: 'all .3s ease-in-out',
+    color: 'white',
     '&:hover': {
       transform: 'scale(1.1)',
+      backgroundColor: 'transparent',
+    },
+    margin: theme.spacing(0, 1, 0, 1),
+    [theme.breakpoints.up('sm')]: {
+      margin: theme.spacing(0, 5, 0, 5),
     },
   },
-};
+  icons: {
+    fontSize: '35px',
+  },
+}));
 
 const NavBar = props => {
-  const { setToggleValue, classes } = props;
-
+  const { setMoreToogle, setToggleValue } = props;
+  const classes = useStyles();
   return (
+    // added to buttons when:
+    // use route to a page: component={Link} to="/foo"
+    // trigger a componet: onClick and redux
     <div>
-      <BottomNavigation className={classes.root} showLabels>
-        <BottomNavigationAction
-          component={Link}
-          to="/"
-          label="Log"
-          value="log"
-          classes={{
-            root: classes.actionItem,
-            selected: classes.selected,
-          }}
-          onClick={() => setToggleValue('myHealthLog')}
-          icon={<FeaturedPlayListOutlinedIcon className={classes.buttons} />}
-        />
-        <BottomNavigationAction
-          component={Link}
-          to="/education"
-          label="Education"
-          value="education"
-          classes={{
-            root: classes.actionItem,
-            selected: classes.selected,
-          }}
-          icon={<StarBorderIcon className={classes.buttons} />}
-        />
-        {/* <BottomNavigationAction
-          label="Chat"
-          value="chat"
-          classes={{
-            root: classes.actionItem,
-            selected: classes.selected,
-          }}
-          icon={<VideocamOutlinedIcon className={classes.buttons} />}
-        /> */}
-        <BottomNavigationAction
-          component={Link}
-          to="/map"
-          label="Map"
-          value="map"
-          classes={{
-            root: classes.actionItem,
-            selected: classes.selected,
-          }}
-          icon={<LocationOnOutlinedIcon className={classes.buttons} />}
-        />
+      <AppBar position="fixed" color="primary" className={classes.appBar}>
+        <Toolbar centered className={classes.toolBar}>
+          <Button component={Link} to="/" className={classes.buttons} onClick={() => setToggleValue('myHealthLog')}>
+            <Grid container direction="column" alignItems="center">
+              <FeaturedPlayListOutlinedIcon className={classes.icons} />
+              <Typography variant="caption">Log</Typography>
+            </Grid>
+          </Button>
+          <Button component={Link} to="/education" className={classes.buttons}>
+            <Grid container direction="column" alignItems="center">
+              <StarBorderIcon className={classes.icons} />
+              <Typography variant="caption">education</Typography>
+            </Grid>
+          </Button>
+          <Button component={Link} to="/map" className={classes.buttons}>
+            <Grid container direction="column" alignItems="center">
+              <LocationOnOutlinedIcon className={classes.icons} />
+              <Typography variant="caption">map</Typography>
+            </Grid>
+          </Button>
+          <Button onClick={() => setMoreToogle(true)} className={classes.buttons}>
+            <Grid container direction="column" alignItems="center">
+              <MoreHorizIcon className={classes.icons} />
+              <Typography variant="caption">more</Typography>
+            </Grid>
+          </Button>
+        </Toolbar>
         <More />
-      </BottomNavigation>
+      </AppBar>
     </div>
   );
 };
-
 NavBar.propTypes = {
-  classes: PropTypes.objectOf(Object).isRequired,
+  setMoreToogle: PropTypes.func.isRequired,
   setToggleValue: PropTypes.func.isRequired,
 };
 const mapDispatchToProps = dispatch => {
   return {
+    setMoreToogle: moreToggel => dispatch(actions.setMoreToogle(moreToggel)),
     setToggleValue: toggleValue => dispatch(actions.setToggleValue(toggleValue)),
   };
 };
 
-export default connect(null, mapDispatchToProps)(withStyles(styles)(NavBar));
+export default connect(null, mapDispatchToProps)(NavBar);
