+++ client/src/redux/actions/navToggle.js
@@ -1,0 +1,8 @@
+export const SET_MORE_TOGGLE = 'SET_MORE_TOGGLE';
+
+export function setMoreToogle(moreToggel) {
+  return {
+    type: SET_MORE_TOGGLE,
+    moreToggel,
+  };
+}
