--- client/src/tests/Slider.test.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import AppSlider from '../components/_archivedComponets/Slider';
-
-test('renders Slider component', () => {
-  render(<AppSlider />);
-});
