+++ client/src/tests/redux/reducers/healthToggle.test.js
@@ -1,6 +1,6 @@
 import expect from 'expect';
 import reducer from 'redux/reducers/healthToggle';
-import { SET_DETAIL_DATA, DELETE_DETAIL_DATA } from '../../../redux/actions/healthToggle';
+import { SET_DETAIL_DATA, DELETE_DETAIL_DATA } from 'redux/actions/healthToggle';
 
 describe('health toggle reducer', () => {
   it('should return the initial state', () => {
